+++ test/lightning_web/live/project_live_test.exs
@@ -18,9 +18,16 @@ defmodule LightningWeb.ProjectLiveTest do
   describe "Index as a regular user" do
     setup :register_and_log_in_user
 
-    test "cannot access the users page", %{conn: conn} do
+    test "cannot access the index page", %{conn: conn} do
       {:ok, _index_live, html} =
-        live(conn, Routes.project_index_path(conn, :index))
+        live(conn, ~p"/settings/projects") |> follow_redirect(conn, "/")
+
+      assert html =~ "You can&#39;t access that page"
+    end
+
+    test "cannot access the new page", %{conn: conn} do
+      {:ok, _index_live, html} =
+        live(conn, ~p"/settings/projects/new")
         |> follow_redirect(conn, "/")
 
       assert html =~ "You can&#39;t access that page"
@@ -100,6 +107,28 @@ defmodule LightningWeb.ProjectLiveTest do
 
       {:ok, view, _html} = live(conn, Routes.project_index_path(conn, :index))
       assert view |> element(delete_button) |> has_element?()
+    end
+
+    test "Edits a project", %{conn: conn} do
+      user = user_fixture()
+      project = project_fixture()
+
+      {:ok, view, _html} = live(conn, ~p"/settings/projects/#{project.id}")
+
+      view
+      |> element("#member_list")
+      |> render_hook("select_item", %{"id" => user.id})
+
+      assert view
+             |> element("button", "Add")
+             |> render_click() =~ "editor"
+
+      view
+      |> form("#project-form")
+      |> render_submit()
+
+      assert_patch(view, ~p"/settings/projects")
+      assert render(view) =~ "Project updated successfully"
     end
   end